refactor(DAO-2135): move dashboard components into subfolder
Move BtcVaultDashboard, BtcVaultMetrics, and BtcVaultActions (plus
tests) into dashboard/ subfolder. Update modal import paths in
BtcVaultActions and page-level mock paths. Add barrel index.ts.
