+++ src/app/btc-vault/BtcVaultPage.tsx
@@ -6,9 +6,9 @@ import { SectionContainer } from '@/app/communities/components/SectionContainer'
 
 import { ActiveRequestSection } from './components/request-processing/ActiveRequestSection'
 import { BtcVaultBanners } from './BtcVaultBanners'
-import { BtcVaultDashboard } from './components/BtcVaultDashboard'
-import { BtcVaultMetrics } from './components/BtcVaultMetrics'
 import { CapitalAllocationSection } from './components/capital-allocation/CapitalAllocationSection'
+import { BtcVaultDashboard } from './components/dashboard/BtcVaultDashboard'
+import { BtcVaultMetrics } from './components/dashboard/BtcVaultMetrics'
 import { BtcVaultDisclosureSection } from './components/disclosure/BtcVaultDisclosureSection'
 import { BtcVaultWalletDisconnectedSection } from './components/disclosure/BtcVaultWalletDisconnectedSection'
 import { useActiveRequests } from './hooks/useActiveRequests'
