+++ src/app/btc-vault/components/dashboard/BtcVaultActions.tsx
@@ -7,15 +7,15 @@ import { Button } from '@/components/Button'
 import { Tooltip } from '@/components/Tooltip'
 import { executeTxFlow } from '@/shared/notification'
 
-import { useActionEligibility } from '../hooks/useActionEligibility'
-import { useBtcVaultInvalidation } from '../hooks/useBtcVaultInvalidation'
-import { useBtcVaultSharesAllowance } from '../hooks/useBtcVaultSharesAllowance'
-import { useSubmitDeposit } from '../hooks/useSubmitDeposit'
-import { useSubmitWithdrawal } from '../hooks/useSubmitWithdrawal'
-import { REQUEST_SUBMITTING_REASON } from '../services/constants'
-import type { DepositRequestParams } from '../services/types'
-import { BtcDepositModal } from './BtcDepositModal'
-import { BtcWithdrawModal } from './BtcWithdrawModal'
+import { useActionEligibility } from '../../hooks/useActionEligibility'
+import { useBtcVaultInvalidation } from '../../hooks/useBtcVaultInvalidation'
+import { useBtcVaultSharesAllowance } from '../../hooks/useBtcVaultSharesAllowance'
+import { useSubmitDeposit } from '../../hooks/useSubmitDeposit'
+import { useSubmitWithdrawal } from '../../hooks/useSubmitWithdrawal'
+import { REQUEST_SUBMITTING_REASON } from '../../services/constants'
+import type { DepositRequestParams } from '../../services/types'
+import { BtcDepositModal } from '../modals/deposit/BtcDepositModal'
+import { BtcWithdrawModal } from '../modals/withdraw/BtcWithdrawModal'
 
 interface BtcVaultActionsProps {
   onRequestSubmitted?: () => void
