+++ src/app/btc-vault/components/dashboard/BtcVaultMetrics.tsx
@@ -12,10 +12,10 @@ import { formatCurrencyWithLabel } from '@/lib/utils'
 import { btcVaultDepositHistory } from '@/shared/constants/routes'
 import { usePricesContext } from '@/shared/context'
 
-import { useEpochState } from '../hooks/useEpochState'
-import { useVaultMetrics } from '../hooks/useVaultMetrics'
-import { formatDateClosingOn, formatDateMonthFirst } from '../services/ui/formatters'
-import { lockedSharePriceToNavPerHumanShareWei } from '../services/vaultShareNav'
+import { useEpochState } from '../../hooks/useEpochState'
+import { useVaultMetrics } from '../../hooks/useVaultMetrics'
+import { formatDateClosingOn, formatDateMonthFirst } from '../../services/ui/formatters'
+import { lockedSharePriceToNavPerHumanShareWei } from '../../services/vaultShareNav'
 
 const PLACEHOLDER = '—'
 const APY_TOOLTIP = 'Annual Percentage Yield — the annualized return on deposited rBTC'
