+++ src/app/btc-vault/components/dashboard/index.ts
@@ -1,0 +1,3 @@
+export { BtcVaultActions } from './BtcVaultActions'
+export { BtcVaultDashboard } from './BtcVaultDashboard'
+export { BtcVaultMetrics } from './BtcVaultMetrics'
